+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.32.0 (2026-08-06)
+
+* feat(CO-4077): prevent duplicate notifications when accepting a proposed new time (#826) ([d851251](https://github.com/zextras/carbonio-calendars-ui/commit/d851251)), closes [#826](https://github.com/zextras/carbonio-calendars-ui/issues/826)
+
+## <small>1.31.1 (2026-08-06)</small>
+
+* fix: use locale language in the propose new time message (#827) ([d3942d7](https://github.com/zextras/carbonio-calendars-ui/commit/d3942d7)), closes [#827](https://github.com/zextras/carbonio-calendars-ui/issues/827)
+
+## 1.31.0 (2026-08-03)
+
+* feat(CO-3662): show original time alongside proposed new time in counter proposals (#821) ([dffab4a](https://github.com/zextras/carbonio-calendars-ui/commit/dffab4a)), closes [#821](https://github.com/zextras/carbonio-calendars-ui/issues/821)
+* fix(CO-3619): include subject and description when forwarding an appointment (#820) ([abdeaaa](https://github.com/zextras/carbonio-calendars-ui/commit/abdeaaa)), closes [#820](https://github.com/zextras/carbonio-calendars-ui/issues/820)
+
 ## <small>1.30.2 (2026-07-31)</small>
 
 * fix(CO-3619): speed up slow view-appointments integration test (#824) ([07395dc](https://github.com/zextras/carbonio-calendars-ui/commit/07395dc)), closes [#824](https://github.com/zextras/carbonio-calendars-ui/issues/824)